Add umbrella application for the IXP topology example.
This application uses the umbrella mac encoding for
the segment routing approach. It considers topologies
that consists of edge switches connected through single hop
cores.

The purpose of this implementation is to help devs to
understand how is the umbrella MAC encoding scheme
and also  to be a starting point for the integration
with iSDX.