Add Umbrella app for iSDX global configuration.
iSDX uses standalone apps to install flows. These apps
communicate with the OpenFlow controller following
a server client model.

The main points implemented are:
    - Parsing of an iSDX like configuration file.
    - Creation of umbrella flows.
    - The idea of two new table identifiers for iSDX.
      (umbrella-edge and umbrella-core).
    - Initial abstract class for load balancing.
    - A dummy load balancing implementation.

The load balancing is just a sketch and still need some
decisions about the implementation design (e.g: how to work
groups?)

The code depends on some iSDX modules. For now it is considering
the iSDX folder in the same level of endeavour.

The code is able to comunicate and send the flow request to the iSDX
reference monitor(basically, it is the iSDX app for the Ryu controller)
Thus, it can be considered a first step towards the integration with iSDX.