Add match on metadata for umbrella.
In order to comply with the load balancing decision,
umbrella matches on the metadata field written on
the load balance table. The metadata field written
is exactly the datapath id of the next hop.

The way load balancing is done now increases
the number of flows in umbrella by the number
of connections from an edge to the core switches.
We might need to think in a more efficient approach
for load balancing.

Also, there is an addition of a flow in the load
balancing table to handle Virtual Next Hops ARPs.
General ARPs are sent directly to the umbrella table, not being
part of the load balance. However, gratuituous ARPs
have virtual MACs and IPs. As the umbrella table has flows to
match on non virtual fields, these fields are translated to real ones
and the packet goes to the umbrella table. Since the load balance
requires a metadata field that works only for ip packets, there
is need to add a flow that writes the metadata field for
gratuitous ARPS.

Later, we should consider adding a group table
for fast failover.
 The virtual MAC is set to the real MAC,